fix: Atualizar variáveis de ambiente, dependências e adicionar TokenProvider, ValidToken e GlobalTokenService. Melhorar estrutura do código, tratamento de erros e segurança.
- Atualizar variáveis de ambiente: adicionar JWT_SECRET_INVITE e configurar JWT_EXPIRES_IN.
- Atualizar dependências e adicionar @nestjs/swagger, @microsoft/tsdoc, class-transformer, class-validator, e tipos do jsonwebtoken.
- Adicionar nova classe TokenProvider em InviteToken.ts para gerar e gerenciar tokens de ID, incluindo criação de payload e verificação de expiração.
- Adicionar nova classe `ValidToken` em `src/Utils/validToken.ts` para validar e decodificar tokens JWT com tratamento de erros personalizado.
- Atualizar AppModule para importar SharedModule e alterar a rota do CheckUserMiddleware para incluir o caminho 'user'.
- Atualizar AuthController: Adicionar tags Swagger API, modificar LoginDTO, CreateAccountDTO, e atualizar os métodos signIn e signUp para usar novos DTOs e melhorar a estrutura do código.
- Atualizar AuthGuard para usar o novo jwtService e a função extractTokenFromHeader.
- Atualizar configuração do AuthModule para incluir signOptions no JwtModule.
- Adicionar JWT_SECRET_INVITE à configuração do ambiente.
- Atualizar filtro de tratamento de exceções para lidar com erros como expiração de token JWT, erros de validação Zod e erros do TypeORM, além de melhorar os logs e mensagens de erro em modo de desenvolvimento.
- Adicionar documentação Swagger e melhorias no tratamento de erros para a API NestJS com suporte à autenticação bearer.
- Atualizar o middleware CheckUser para usar o GlobalTokenService e lidar com autorização em várias rotas.
- Adicionar opções de configuração de pipes de validação com importações para interfaces externas do NestJS.
- Atualizar e refatorar o controlador de posts para melhor organização do código e segurança, integrando autenticação e autorização usando o GlobalTokenService e impondo restrições de acesso com base na escola e no papel do usuário.
- Atualizar `posts.module.ts` para incluir `SharedModule` nas importações e adicionar `PostsService` nas exportações.
- Atualizar e refatorar controlador e serviços relacionados a escolas, incluindo validação, autenticação e gerenciamento de tokens.
- Atualizar importações do módulo de escolas e adicionar UserSchoolAssociationModule e SharedModule.
- Atualizar AuthService para lidar com conflitos e melhorar os processos de autenticação e cadastro de usuários.
- Atualizar arquivo post.service.ts para adicionar novos parâmetros, melhorar funções de consulta e tratar exceções para posts ausentes.
- Atualizar school.service.ts: Adicionar tratamento de NotFoundException para ID de escola ausente no método findById e atualizar o parâmetro do método delete de number para string.
- Atualizar user.service.ts: Adicionar NotFoundException para melhor tratamento de erros e corrigir o método findUserByEmail para retornar o usuário correto.
- Atualizar serviço de UserSchoolAssociation para suportar busca por múltiplos IDs de escola e tipo de usuário, além de melhorar o tratamento de exceções em várias funções.
- Adicionar classe GlobalTokenService com métodos set e get para gerenciar o token decodificado.
- Criar um novo módulo compartilhado em src/shared/shared.module.ts, importando GlobalTokenService e exportando para uso em outros módulos.
- Atualizar e refatorar UserSchoolAssociationController para usar novos DTOs, melhorar o tratamento de erros e adicionar nova funcionalidade para convidar usuários para uma escola e verificar tokens.
- Atualizar importações e dependências de módulos do UserSchoolAssociationModule para melhorar a organização do código e integração com outros módulos.
- Atualizar controlador de usuários com melhor organização do código, uso de DTOs e implementação de segurança de API e autenticação.
- Atualizar `users.module.ts`: Adicionar `SharedModule` nas importações e incluí-lo no array de importações.

+++ src/posts/controller/post.controller.ts
@@ -1,6 +1,5 @@
 import { IPost } from '@/entities/interfaces/posts.interface';
 import { PostsService } from '@/services/post.service';
-import { ZodValidationPipe } from '@/pipe/zod-validation.pipe';
 import {
   Body,
   Controller,
@@ -11,86 +10,211 @@ import {
   Post,
   Put,
   Query,
+  Req,
   Res,
   UseGuards,
 } from '@nestjs/common';
-import { Response } from 'express';
+import { response, Response } from 'express';
 import { AuthGuard } from '@/auth/auth.guard';
-import { CreatePost, createPostScheme } from '../TypesRequestZod/CreatePost.Type';
-import { UpdatePost, updatePostScheme } from '../TypesRequestZod/UpdatePost.Type';
-import { GetPostById, getPostByIdScheme } from '../TypesRequestZod/GetPostById.Type';
-import { GetAllPosts, getAllPosts } from '../TypesRequestZod/GetAllPost.Type';
-import { GetPostByKeyWord, getPostByKeyWord } from '../TypesRequestZod/GetPostByKeyWord';
+
 import { UserSchoolAssociationService } from '@/services/userSchoolAssociation.service';
 import { TypeUser } from '@/entities/models/userSchoolAssociation.entity';
+import { ApiTags } from '@nestjs/swagger';
+import { GetAllPostDTO } from '../DTOS/getAllPost.dto';
+import { GetPostByIdDTO } from '../DTOS/getPostById.dto';
+import { GetPostByKeyWordDTO } from '../DTOS/getPostByKeyWord.dto';
+import { CreatePostDTO } from '../DTOS/createPost.dto';
+import { UpdatePostDTO } from '../DTOS/updatePost.dto';
+import { headerDTO } from '@/Utils/dtos/header.dto';
+import { GlobalTokenService } from '@/shared/globalTokenService';
 
+@ApiTags('Post')
 @UseGuards(AuthGuard)
 @Controller('posts')
 export class PostController {
   constructor(
     private userSchoolAssociationService: UserSchoolAssociationService,
     private postsService: PostsService,
+    private readonly globalTokenService: GlobalTokenService,
   ) {}
 
   @Get()
   async getAllPosts(
-    @Query(new ZodValidationPipe(getAllPosts)) { page, limit }: GetAllPosts,
+    @Query() { page, limit }: GetAllPostDTO,
+    @Headers() { schoolid }: headerDTO,
+    @Res() response: Response,
   ) {
-    return await this.postsService.findAll(page, limit);
+    const decoded = this.globalTokenService.getDecodedToken();
+
+    if (decoded.schools.filter((x) => x.schoolId == schoolid).length == 0) {
+      return response.status(403).json({
+        message: 'User Not Authorized To get this posts',
+      });
+    }
+
+    const associations =
+      await this.userSchoolAssociationService.findAllBySchoolId(schoolid);
+
+    const ids = associations.map((item) => item.id);
+
+    const result = await this.postsService.findAll(page, limit, ids);
+
+    response.status(200).json(result);
+  }
+
+  @Get('drafts')
+  async getAllPostsDrafts(
+    @Query() { page, limit }: GetAllPostDTO,
+    @Headers() { schoolid }: headerDTO,
+    @Res() response: Response,
+  ) {
+    const decoded = this.globalTokenService.getDecodedToken();
+
+    const school = decoded.schools.find((x) => x.schoolId == schoolid);
+
+    if (!school) {
+      return response.status(403).json({
+        message: 'User Not Authorized To get this posts',
+      });
+    }
+
+    if (school.typeUser != TypeUser.Professor) {
+      return response.status(403).json({
+        message: 'User Not Authorized To get this posts',
+      });
+    }
+
+    const associations =
+      await this.userSchoolAssociationService.findAllBySchoolId(schoolid);
+
+    const ids = associations.map((item) => item.id);
+
+    const result = await this.postsService.findDraftsByAssociationId(
+      ids,
+      page,
+      limit,
+    );
+
+    response.status(200).json(result);
   }
 
   @Get(':id')
   async getPostById(
-    @Param(new ZodValidationPipe(getPostByIdScheme)) { id }: GetPostById,
+    @Param() { id }: GetPostByIdDTO,
+    @Headers() { schoolid }: headerDTO,
+    @Res() response: Response,
   ) {
-    return await this.postsService.findById(id);
-  }
+    const decoded = this.globalTokenService.getDecodedToken();
 
-  @Get('school/:id')
-  async getPostBySchoolId(
-    @Param('id') id: string,
-    @Query(new ZodValidationPipe(getAllPosts)) { page, limit }: GetAllPosts,
-  ) {
-    const result = await this.userSchoolAssociationService.findAllBySchoolIdAndTypeUser(id, TypeUser.Professor);
+    if (decoded.schools.filter((x) => x.schoolId == schoolid).length == 0) {
+      return response.status(403).json({
+        message: 'User Not Authorized To get this posts',
+      });
+    }
 
-    const getAssociationsIds = result.map((item) => item.id);
+    const school = decoded.schools.find((x) => x.schoolId == schoolid);
 
-    const posts = await this.postsService.findPostByAssociationId(getAssociationsIds, page, limit);
+    const result = await this.postsService.findById(id);
 
-    return posts;
-  }
+    if (result.isDraft) {
+      if (school.typeUser != TypeUser.Professor) {
+        return response.status(403).json({
+          message: 'User Not Authorized To get this posts',
+        });
+      }
+    }
 
-  @Get('school/draft/:id')
-  async getDraftPostBySchoolId(
-    @Param('id') id: string,
-    @Query(new ZodValidationPipe(getAllPosts)) { page, limit }: GetAllPosts,
-  ) {
-    const result = await this.userSchoolAssociationService.findAllBySchoolIdAndTypeUser(id, TypeUser.Professor);
-
-    const getAssociationsIds = result.map((item) => item.id);
-
-    return await this.postsService.findDraftsByAssociationId(getAssociationsIds, page, limit);
+    response.status(200).json(result);
   }
 
   @Get('find/school')
-  async getPostByKeyWord(@Query(new ZodValidationPipe(getPostByKeyWord)) { search }: GetPostByKeyWord,
-    @Query(new ZodValidationPipe(getAllPosts)) { page, limit }: GetAllPosts,
-    @Headers('X-SchoolId') id: string) {
+  async getPostByKeyWord(
+    @Query()
+    { search }: GetPostByKeyWordDTO,
+    @Query() { page, limit }: GetAllPostDTO,
+    @Headers() { schoolid }: headerDTO,
+  ) {
+    const decoded = this.globalTokenService.getDecodedToken();
 
-    const result = await this.userSchoolAssociationService.findAllBySchoolIdAndTypeUser(id, TypeUser.Professor);
+    if (decoded.schools.filter((x) => x.schoolId == schoolid).length == 0) {
+      return response.status(403).json({
+        message: 'User Not Authorized To get this posts',
+      });
+    }
+
+    const result =
+      await this.userSchoolAssociationService.findAllBySchoolId(schoolid);
 
     const getAssociationsIds = result.map((item) => item.id);
 
-    return await this.postsService.findPostByKeyWord(getAssociationsIds, search, page, limit);
+    return await this.postsService.findPostByKeyWord(
+      getAssociationsIds,
+      search,
+      page,
+      limit,
+    );
+  }
+
+  @Get('finddraft/school')
+  async getPostDraftByKeyWord(
+    @Query()
+    { search }: GetPostByKeyWordDTO,
+    @Query() { page, limit }: GetAllPostDTO,
+    @Headers() { schoolid }: headerDTO,
+  ) {
+    const decoded = this.globalTokenService.getDecodedToken();
+
+    const school = decoded.schools.find((x) => x.schoolId == schoolid);
+
+    if (!school) {
+      return response.status(403).json({
+        message: 'User Not Authorized To get this posts',
+      });
+    }
+
+    if (school.typeUser != TypeUser.Professor) {
+      return response.status(403).json({
+        message: 'User Not Authorized To get this posts',
+      });
+    }
+    const result =
+      await this.userSchoolAssociationService.findAllBySchoolId(schoolid);
+
+    const getAssociationsIds = result.map((item) => item.id);
+
+    return await this.postsService.findPostDraftsByKeyWord(
+      getAssociationsIds,
+      search,
+      page,
+      limit,
+    );
   }
 
   @Post()
   async createPost(
-    @Body(new ZodValidationPipe(createPostScheme))
-    { title, content, isDraft,status,associationSchool }: CreatePost,
+    @Body()
+    { title, content, isDraft, status, associationSchool }: CreatePostDTO,
     @Res() response: Response,
+    @Headers() { schoolid }: headerDTO,
   ) {
-    const userSchoolAssociation = await this.userSchoolAssociationService.findById(associationSchool);
+    const decoded = this.globalTokenService.getDecodedToken();
+
+    const school = decoded.schools.find((x) => x.schoolId == schoolid);
+
+    if (!school) {
+      return response.status(403).json({
+        message: 'User Not Authorized To Create Posts For this School',
+      });
+    }
+
+    if (school.typeUser != TypeUser.Professor) {
+      return response.status(403).json({
+        message: 'User Not Authorized To Create posts',
+      });
+    }
+
+    const userSchoolAssociation =
+      await this.userSchoolAssociationService.findById(associationSchool);
 
     if (!userSchoolAssociation) {
       return response.status(404).json('User School Association not found');
@@ -101,7 +225,7 @@ export class PostController {
       content,
       isDraft,
       status,
-      userSchoolAssociation
+      userSchoolAssociation,
     });
 
     return response.status(201).send(result);
@@ -109,11 +233,28 @@ export class PostController {
 
   @Put(':id')
   async updatePost(
-    @Param(new ZodValidationPipe(getPostByIdScheme)) { id }: GetPostById,
-    @Body(new ZodValidationPipe(updatePostScheme))
-    { title, content, isDraft, status }: UpdatePost,
+    @Param() { id }: GetPostByIdDTO,
+    @Body()
+    { title, content, isDraft, status }: UpdatePostDTO,
     @Res() response: Response,
+    @Headers() { schoolid }: headerDTO,
   ) {
+    const decoded = this.globalTokenService.getDecodedToken();
+
+    const school = decoded.schools.find((x) => x.schoolId == schoolid);
+
+    if (!school) {
+      return response.status(403).json({
+        message: 'User Not Authorized To Update Posts For this School',
+      });
+    }
+
+    if (school.typeUser != TypeUser.Professor) {
+      return response.status(403).json({
+        message: 'User Not Authorized To Update posts',
+      });
+    }
+
     let post = (await this.postsService.findById(id)) as IPost;
 
     if (!post) {
@@ -131,7 +272,27 @@ export class PostController {
   }
 
   @Delete(':id')
-  async deletePost(@Param('id') id: number, @Res() response: Response) {
+  async deletePost(
+    @Param('id') id: number,
+    @Res() response: Response,
+    @Headers() { schoolid }: headerDTO,
+  ) {
+    const decoded = this.globalTokenService.getDecodedToken();
+
+    const school = decoded.schools.find((x) => x.schoolId == schoolid);
+
+    if (!school) {
+      return response.status(403).json({
+        message: 'User Not Authorized To Delete Posts For this School',
+      });
+    }
+
+    if (school.typeUser != TypeUser.Professor) {
+      return response.status(403).json({
+        message: 'User Not Authorized To Delete posts',
+      });
+    }
+
     await this.postsService.delete(id);
 
     return response.status(204).send();
